pyproject.toml
[project]
authors = [{name = "jcav", email = "micojcav@outlook.com"}]
name = "alpenstock"
requires-python = ">= 3.11"
version = "1.2.1"
readme = "README.md"
description = "A collection of handy Python utilities designed to help developers tackle the challenges of Python programming."

dependencies = [
    "loguru>=0.7.0",
    "pydantic>=2.3.0",
    "attrs",
    "jaxtyping",
    "wadler-lindig",
    "ruamel-yaml>=0.18",
]

[project.urls]
Homepage = "https://github.com/gjcav/alpenstock/"
Documentation = "https://gjcav.github.io/alpenstock/"
Repository = "https://github.com/gjcav/alpenstock/"

[project.optional-dependencies]
dev = [
    "ipykernel", 
    "lmdb",
    "numpy",
    "torch", 
    "pytest",
    "pyright==1.1.350",
    "mkdocs-material",
    "mkdocstrings[python]",
    "mkdocs-jupyter",
    "mkdocs-gen-files",
    "mkdocs-awesome-nav",
    "mkdocs-glightbox",
    "mkdocs-api-autonav",
    "griffe-pydantic",
    "griffe-warnings-deprecated",
    "build",
    "twine"
]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64", "win-64"]

[tool.pixi.pypi-dependencies]
alpenstock = { path = ".", editable = true }

[tool.pixi.environments]
default = { features=["dev"], solve-group = "default" }

[tool.pixi.tasks]
